Split test_product_media_create.py file.
